[local] [whisk] Unify storage and resource handling
